extract kinds directory
