--- bazel/BUILD.bazel
@@ -1,11 +1,0 @@
-alias(
-    name = "rules",
-    actual = "//bazel/src/main/kotlin/org/virtuslab/bazelsteward/rules",
-    visibility = ["//visibility:public"],
-)
-
-alias(
-    name = "bazel",
-    actual = "//bazel/src/main/kotlin/org/virtuslab/bazelsteward/bazel",
-    visibility = ["//visibility:public"],
-)
