--- bazel/src/main/kotlin/org/virtuslab/bazelsteward/bazel/BUILD.bazel
@@ -1,12 +1,0 @@
-library(
-    name = "bazel",
-    srcs = glob(["**/*.kt"]),
-    resources = glob(
-        ["resources/**"],
-    ),
-    deps = [
-        "//core",
-        "@maven//:com_fasterxml_jackson_core_jackson_core",
-        "@maven//:com_fasterxml_jackson_module_jackson_module_kotlin",
-    ],
-)
