--- bazel/src/main/kotlin/org/virtuslab/bazelsteward/rules/BUILD.bazel
@@ -1,13 +1,0 @@
-library(
-    name = "rules",
-    srcs = glob(["**/*.kt"]),
-    resources = ["//bazel/resources:rules_resources"],
-    deps = [
-        "//core",
-        "@maven//:com_fasterxml_jackson_dataformat_jackson_dataformat_yaml",
-        "@maven//:com_fasterxml_jackson_module_jackson_module_kotlin",
-        "@maven//:com_networknt_json_schema_validator",
-        "@maven//:org_jetbrains_kotlinx_kotlinx_cli_jvm",
-        "@maven//:org_json_json_20220924",
-    ],
-)
