--- bazel/src/test/kotlin/org/virtuslab/bazelsteward/bazel/BUILD.bazel
@@ -1,7 +1,0 @@
-unit_tests(
-    srcs = glob(["**/*.kt"]),
-    test_package = "org.virtuslab.bazelsteward.bazel",
-    deps = [
-        "//bazel",
-    ],
-)
