+++ kinds/bazel/resources/BUILD.bazel
