+++ kinds/bazel/resources/dump_repositories.bzlignore
