+++ kinds/bazel/src/main/kotlin/org/virtuslab/bazelsteward/bazel/version/BUILD.bazel
@@ -1,0 +1,12 @@
+library(
+    name = "version",
+    srcs = glob(["**/*.kt"]),
+    resources = glob(
+        ["resources/**"],
+    ),
+    deps = [
+        "//core",
+        "@maven//:com_fasterxml_jackson_core_jackson_core",
+        "@maven//:com_fasterxml_jackson_module_jackson_module_kotlin",
+    ],
+)
