+++ kinds/bazel/src/test/kotlin/org/virtuslab/bazelsteward/bazel/rules/BUILD.bazel
@@ -1,0 +1,7 @@
+unit_tests(
+    srcs = glob(["**/*.kt"]),
+    test_package = "org.virtuslab.bazelsteward.bazel.rules",
+    deps = [
+        "//kinds/bazel:rules",
+    ],
+)
