+++ kinds/bazel/src/test/kotlin/org/virtuslab/bazelsteward/bazel/rules/BazelRuleLibraryIdTest.kt
