+++ kinds/bazel/src/test/kotlin/org/virtuslab/bazelsteward/bazel/version/BUILD.bazel
@@ -1,0 +1,7 @@
+unit_tests(
+    srcs = glob(["**/*.kt"]),
+    test_package = "org.virtuslab.bazelsteward.bazel.version",
+    deps = [
+        "//kinds/bazel:version",
+    ],
+)
