+++ kinds/maven/BUILD.bazel
@@ -1,0 +1,5 @@
+alias(
+    name = "maven",
+    actual = "//kinds/maven/src/main/kotlin/org/virtuslab/bazelsteward/maven",
+    visibility = ["//visibility:public"],
+)
