+++ kinds/maven/src/main/kotlin/org/virtuslab/bazelsteward/maven/BUILD.bazel
