+++ kinds/maven/src/main/kotlin/org/virtuslab/bazelsteward/maven/MavenRepository.kt
