+++ kinds/maven/src/test/kotlin/org/virtuslab/bazelsteward/maven/BUILD.bazel
@@ -1,0 +1,7 @@
+unit_tests(
+    srcs = glob(["**/*.kt"]),
+    test_package = "org.virtuslab.bazelsteward.maven",
+    deps = [
+        "//kinds/maven/src/main/kotlin/org/virtuslab/bazelsteward/maven",
+    ],
+)
