+++ kinds/maven/src/test/kotlin/org/virtuslab/bazelsteward/maven/MavenCoordinatesTest.kt
