--- maven/BUILD.bazel
@@ -1,5 +1,0 @@
-alias(
-    name = "maven",
-    actual = "//maven/src/main/kotlin/org/virtuslab/bazelsteward/maven",
-    visibility = ["//visibility:public"],
-)
